ergo: add lazylock update commit autocmd
When the lazy lock file is updated, a new commit should (hopefully) be
created. This should allow for better versioning. We'll see.

Additionally add empty files under the config directory. I'm not using
these right now, but these are files that lazyvim searches for. Good to
have them so I know where to put things.